Proyecto Final - Thiago Mowszet

+++ Proyecto Final/main.js
@@ -26,14 +26,44 @@ class TurnList {
         this.tablebody = this.table.appendChild(document.createElement("tbody"));
         let ref = this;
         this.addbutton.addEventListener("click", () => {
-            let dr = prompt("Seleccione la especialidad del dr: (Ej: Traumatologo)");
-            let dayH = prompt("Seleccione el dia que desee atenderse: (Ej: 25 de Abril)");
-            let hourH = prompt("Seleccione la hora: (Ej: 22hrs)");
-            let turn = document.createElement("tr");
-            turn.innerHTML = `<td>${dr}</td><td>${dayH}</td><td>${hourH}</td>`;
-            ref.tablebody.append(turn);
-            this.list.push(new Turn(dr, dayH, hourH));
-            sessionStorage.setItem("turns", JSON.stringify(this.list));
+            Swal.fire({
+                preConfirm: (response) => {
+                    let especialidad = document.getElementById("especialidad").value
+                    let diaTurno = document.getElementById("diaTurno").value
+                    let horaTurno = document.getElementById("horaTurno").value
+
+                    if (response) {
+                        if (especialidad == "" || diaTurno == "" || horaTurno == "") {
+                            Swal.fire({
+                                icon: "error",
+                                title: "Error",
+                                text: "Debe de llenar todos los campos"
+                            })
+                        } else {
+                            let turn = document.createElement("tr")
+                            turn.innerHTML = `<td>${especialidad}</td><td>${diaTurno}</td><td>${horaTurno}</td>`;
+                            ref.tablebody.append(turn);
+                            this.list.push(new Turn(especialidad, diaTurno, horaTurno));
+                            sessionStorage.setItem("turns", JSON.stringify(this.list));
+                            Swal.fire({
+                                icon: "success",
+                                title: "Listo",
+                                text: "Se ha registrado su turno"
+                            });
+                        }
+                    }
+                },
+                icon: "question",
+                html: `
+                <div class="addFormAlert">
+                <label for="especialidad">Ingrese la Especialidad</label>
+                <input id="especialidad" type="text">
+                <label for="diaTurno">Ingrese el dia del turno</label>
+                <input id="diaTurno" type="text">
+                <label for="horaTurno">Ingrese la hora del turno</label>
+                <input id="horaTurno" type="text">
+                </div>`
+            });
         });
     }
 }
